# docs/swagger.yaml
components:
  schemas:
...
    github_com_akira-core_kube-state-graph_pkg_graph.EdgeType:
      enum:
      - pod-mounts-pvc
      - pod-calls-pod
      - pod-calls-service
      - service-selects-pod
      - pod-to-node
      - pvc-to-storageclass
      type: string
      x-enum-varnames:
      - EdgeTypePodMountsPVC
      - EdgeTypePodCallsPod
      - EdgeTypePodCallsService
      - EdgeTypeServiceSelectsPod
      - EdgeTypePodToNode
      - EdgeTypePVCToStorageClass
    github_com_akira-core_kube-state-graph_pkg_graph.EdgeTypeDefinition:
      properties:
        description:
          type: string
        directed:
          type: boolean
        labels:
          items:
            $ref: '#/components/schemas/github_com_akira-core_kube-state-graph_pkg_graph.EdgeTypeLabel'
          type: array
          uniqueItems: false
        may_cross_cluster:
          type: boolean
        source_type:
          items:
            $ref: '#/components/schemas/github_com_akira-core_kube-state-graph_pkg_graph.NodeType'
          type: array
          uniqueItems: false
        target_type:
          items:
            $ref: '#/components/schemas/github_com_akira-core_kube-state-graph_pkg_graph.NodeType'
          type: array
          uniqueItems: false
        type:
          $ref: '#/components/schemas/github_com_akira-core_kube-state-graph_pkg_graph.EdgeType'
      type: object
    github_com_akira-core_kube-state-graph_pkg_graph.EdgeTypeLabel:
      properties:
        description:
          type: string
        name:
          type: string
        value_type:
          type: string
      type: object
    github_com_akira-core_kube-state-graph_pkg_graph.NodeType:
      enum:
      - pod
      - node
      - pvc
      - service
      - external
      - storageclass
      type: string
      x-enum-varnames:
      - NodeTypePod
      - NodeTypeK8sNode
      - NodeTypePVC
      - NodeTypeService
      - NodeTypeExternal
      - NodeTypeStorageClass
...
    internal_api.ClusterInfo:
      properties:
        name:
          type: string
      type: object
    internal_api.clustersBody:
      properties:
        apiVersion:
          type: string
        clusters:
          items:
            $ref: '#/components/schemas/internal_api.ClusterInfo'
          type: array
          uniqueItems: false
      type: object
    internal_api.edgeTypesBody:
      properties:
        apiVersion:
          type: string
        edge_types:
          items:
            $ref: '#/components/schemas/github_com_akira-core_kube-state-graph_pkg_graph.EdgeTypeDefinition'
          type: array
          uniqueItems: false
      type: object
    internal_api.errorBody:
      properties:
        apiVersion:
          type: string
        error:
          $ref: '#/components/schemas/internal_api.errorField'
      type: object
    internal_api.errorField:
      properties:
        message:
          type: string
        reason:
          type: string
      type: object
  securitySchemes:
    ApiKeyAuth:
      description: API key presented in the `X-API-Key` header. Required on `/v1/*`
        when the server is started with keys configured. Health, metrics, and docs
        routes are exempt.
      in: header
      name: X-API-Key
      type: apiKey
...
info:
  description: |-
    Multi-cluster pod / node / PVC graph API. Reads kube-state-metrics and pod-UID-resolved service-graph metrics from a centralised VictoriaMetrics and returns the joined cross-cluster graph as Cytoscape.js JSON.

    **Authentication.** When the server is started with API keys configured (`--api-keys-file` or `--api-keys`), every request to `/v1/*` MUST carry an `X-API-Key: <key>` header. Missing or invalid keys yield `401 Unauthorized`. Health probes (`/livez`, `/readyz`), the metrics endpoint (`/metrics`), and the OpenAPI / Scalar UI routes (`/openapi.*`, `/docs`) are exempt and require no key.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: kube-state-graph API
  version: v1
openapi: 3.1.0
paths:
...
  /v1/clusters:
    get:
      description: |-
        Returns the set of clusters observed in `kube_node_info` over a fixed 1 h lookback. Each request hits VictoriaMetrics directly under `--api-timeout`.

        <details><summary><b>Sample response</b></summary>

        ```json
        {
        "apiVersion": "v1",
        "clusters": [
        { "name": "prod-eu" },
        { "name": "prod-us" },
        { "name": "stage-eu" }
        ]
        }
        ```

        </details>
      parameters:
      - description: API key. Required when the server is started with API keys configured.
        in: header
        name: X-API-Key
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_api.clustersBody'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_api.errorBody'
          description: Missing or invalid `X-API-Key` (only when API key auth is configured)
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_api.errorBody'
          description: Upstream VictoriaMetrics returned an error
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_api.errorBody'
          description: Discovery query exceeded --api-timeout
      security:
      - ApiKeyAuth: []
      summary: List clusters
      tags:
      - discovery
  /v1/edge-types:
    get:
      description: |-
        Static catalogue of edge types this server can produce — directionality, valid source/target node types, supported labels, and whether the edge may cross cluster boundaries. No upstream calls; served with `Cache-Control: public, max-age=3600`. Use this to validate the `edge_type` filter on `/v1/graph` and to drive UI legends.

        <details><summary><b>Edge type matrix</b></summary>

        | type | source → target | directed | cross-cluster |
        |---|---|---|---|
        | `pod-mounts-pvc` | pod → pvc | yes | no |
        | `pod-calls-pod` | pod \| service \| external → pod \| external | yes | yes |
        | `pod-calls-service` | pod \| service \| external → service | yes | yes |
        | `service-selects-pod` | service → pod | yes | no |

        A call endpoint whose pod UID is empty and whose `client`/`server` label is a `://` connection string is resolved (detection hardcoded, no operator knob) against every loaded cluster in the caller's family (cluster names equal after normalising digit runs; anchored on the UID-recovered client-pod cluster when available, else the trace-source label): each surviving family cluster holding the addressed Service yields a `service` node and a `pod-calls-service` edge (which may therefore cross clusters). Candidates provably without backing pods (zero endpoints in an endpoint-visible cluster) are pruned when an endpoint-backed sibling exists; an anchor naming no loaded family falls back to the single loaded family holding the Service (multi-family names are ambiguous and stay external). Zero surviving candidates yield an `external` node. A non-URL missing-UID label resolves to an `external` node. `service-selects-pod` edges are materialised on demand from `service` nodes to their own cluster's backing pods.

        </details>
      parameters:
      - description: API key. Required when the server is started with API keys configured.
        in: header
        name: X-API-Key
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_api.edgeTypesBody'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_api.errorBody'
          description: Missing or invalid `X-API-Key` (only when API key auth is configured)
      security:
      - ApiKeyAuth: []
      summary: Edge-type catalogue
      tags:
      - discovery
...
servers:
- url: /